server/apiServerForAngular/requirements.txt
Django==1.8.3
djangorestframework==3.1.3
Markdown==2.6.2
django_filter==0.10.0
requests==2.7.0
